+++ doc-user-mkdocs-project/docs/appendix-release-notes/release-notes.md
@@ -4,8 +4,12 @@
 
 ----------
 
-## Changes in Version 1.2.0 ##
+## Changes in Version 1.0.0 ##
 
 **Feature release - initial production release.**
 
-* ![new](new.png) [1.0.0] Initial production release.
+* ![new](new.png) [1.0.0] Initial production release:
+    +   Main TSTool window includes browsing features to list KiWIS time series.
+    +   The [`ReadKiWIS`](../command-ref/ReadKiWIS/ReadKiWIS.md) command is provided to automate
+        reading 1+ time series.
+    +   Documentation is available for the [KiWIS datastore](../datastore-ref/KiWIS/KiWIS.md).